Complete CRUD including Patch for Students

+++ server/app.py
@@ -22,6 +22,8 @@ api = Api(app)
 class Students(Resource):
     def get(self):
         students = Student.query.all()
+        if not students:
+            return make_response({"error": "Empty list"},404)
         students_list = [student.to_dict() for student in students]
         return make_response(students_list,200)
     def post(self):
@@ -35,11 +37,28 @@ class Students(Resource):
 class StudentbyId(Resource):
     def get(self,id):
         student = Student.query.get(id)
+        if not student:
+            return make_response({"error": "Student not found"},404)
         return make_response(student.to_dict(),200)
     def patch(self,id):
-        pass
+        student = Student.query.get(id)
+        if not student:
+            return make_response({"error": "Student not found"},404)
+        try:
+            incoming = request.get_json()
+            for attr in incoming:
+                setattr(student,attr,incoming[attr])
+                db.session.commit()
+                return make_response(student.to_dict(),202)
+        except:
+            return make_response({"errors": ["validation errors"]},400)
+    
+            
+        
     def delete(self,id):
         student = Student.query.get(id)
+        if not student:
+            return make_response({"error": "Student not found"},404)
         db.session.delete(student)
         db.session.commit()
